Crypto.NaCl.Auth.Auth

Description

Fast, cryptographically strong authentication.

Documentation

authenticate

Arguments

:: SecretKey
Secret key

-> ByteString
Message

-> ByteString
Authenticator

verify

Arguments

:: SecretKey
Key

-> ByteString
Authenticator returned via authenticate

-> ByteString
Message

-> Bool
Result: True if properly verified, False otherwise

authKeyLength :: Int

authKeyLength is the required key length for a key given to authenticate or verify. Using any other key length will result in error.