Crypto.NaCl.Encrypt.PublicKey

Description

Public-key encryption.

Types

data PKNonce

Keypair creation

createKeypair :: IO KeyPair

Randomly generate a public and private key for doing authenticated encryption.

Encryption, Decryption

encrypt

Arguments

:: PKNonce
Nonce

-> ByteString
Message

-> PublicKey
Recievers public key

-> SecretKey
Senders secret key

-> ByteString
Ciphertext

decrypt

Arguments

:: PKNonce
Nonce

-> ByteString
Input ciphertext

-> PublicKey
Senders public key

-> SecretKey
Recievers secret key

-> Maybe ByteString
Ciphertext

Precomputation interface

If you send many messages to the same receiver, or receive many messages from the same sender, you can gain speed increases by instead using the following precomputation interface, which splits the encryption and decryption steps into two parts.

For encryption, you first create an NM by using createNM, using the senders secret key, and receivers public key. You can then use encryptNM to encrypt data.

For decryption, you first create an NM by using createNM, using the recievers secret key, and the senders publickey. You can then use decryptNM to decrypt data.

data NM

An NM is intermediate data computed by createNM given a public and private key which can be used to encrypt/decrypt information via encryptNM or decryptNM.

An NM can be re-used between two communicators for any number of messages.

Its name is not particularly enlightening as to its purpose, it is merely the same identifier used in the NaCl source code for this interface.

createNM :: KeyPair -> NM

Creates an intermediate piece of NM data for sending/receiving messages to/from the same person. The resulting NM can be used for any number of messages between client/server.

encryptNM :: NM -> PKNonce -> ByteString -> ByteString

Encrypt data from a specific sender to a specific receiver with some precomputed NM data.

decryptNM :: NM -> PKNonce -> ByteString -> Maybe ByteString

Decrypt data from a specific sender for a specific receiver with some precomputed NM data.

Miscellaneous

publicKeyLength :: Int

Length of a PublicKey in bytes.

secretKeyLength :: Int

Length of a SecretKey in bytes.

nmLength :: Int

Length of the intermediate NM data used by the precomputation interface.