Crypto.NaCl.Encrypt.Stream

Description

Fast stream encryption.

Types

data StreamNonce

Stream generation

streamGen

Arguments

:: StreamNonce
Nonce

-> Int
Size

-> SecretKey
Input

-> ByteString
Resulting crypto stream

Given a Nonce n, size s and SecretKey sk, streamGen n s sk generates a cryptographic stream of length s.

Encryption and decryption

encrypt

Arguments

:: StreamNonce
Nonce

-> ByteString
Input plaintext

-> SecretKey
Secret key

-> ByteString
Ciphertext

Given a Nonce n, plaintext p and SecretKey sk, encrypt n p sk encrypts the message p using SecretKey sk and returns the result.

encrypt guarantees the resulting ciphertext is the plaintext bitwise XOR'd with the result of streamGen. As a result, encrypt can also be used to decrypt messages.

decrypt

Arguments

:: StreamNonce
Nonce

-> ByteString
Input ciphertext

-> SecretKey
Secret key

-> ByteString
Plaintext

Simple alias for encrypt.

Misc

keyLength :: Int

Length of a SecretKey needed for encryption/decryption.