Crypto.NaCl.Nonce

Description

Simple API for cryptographic nonces.

Nonce class

class Nonce k where

Methods

size :: Tagged k Int

toBS :: k -> ByteString

fromBS :: ByteString -> Maybe k

createZeroNonce :: k

Create an empty nonce of length n where all the bytes are zero.

createRandomNonce :: IO k

Create a random nonce, seeded by devurandom.

clearBytes :: Int -> k -> k

clearBytes n nonce clears the last n bytes of the Nonce and makes them all 0. This is useful for the pattern of generating a cryptographic nonce randomly, clearing the last n bytes, and then using incNonce to increment the Nonce for communication with another party.

Invariants:

n must be less than the size of the nonce

Properties:

 clearBytes (nonceLen nonce) nonce == createZeroNonce (nonceLen nonce)

incNonce :: k -> k

Increment a Nonce by 1.