Crypto.NaCl.Sign

Description

This module gives you the ability to create signed messages and verify them against a signer's public key.

Please be cautious using this module. It will change significantly in the next release of NaCl, and will be based on ed25519 (see http://ed25519.cr.yp.to).

Keypair creation

createKeypair :: IO KeyPair

Randomly generate a public and private key for doing authenticated signing and verification.

Signing and verifying messages

sign

Arguments

:: SecretKey
Signers secret key

-> ByteString
Input message

-> ByteString
Resulting signed message

Sign a message with a particular SecretKey.

verify

Arguments

:: PublicKey
Signers public key

-> ByteString
Signed message

-> Maybe ByteString
Verification check

Verifies a signed message with a PublicKey. Returns Nothing if verification fails, or Just xs where xs is the original message if it succeeds.

signPublicKeyLength :: Int

Length of a signers PublicKey, in bytes.

signSecretKeyLength :: Int

Length of a signers SecretKey, in bytes.